requirements/requirements-test.txt
# Tox to make it available in the Docker container.
tox==3.3.0

# Pytest packages.
pytest-django==3.4.2
pytest==3.8.0
pytest-cov==2.6.0
